+++ apps/audience/src/shared/configs/firebase.ts
